Swapping to DOM creation through methods in addItem.
This and previous checkin are largely to move from the nuking I was doing earlier with assigning event handlers to everything matching a className, regardless of if it'd had an event handler assigned earlier or not (we would take each matching element, remove the old handler, if it existed, and replace with a new one), which was insane. Now we explicitly assign event handlers on the elements as we create them, which is "The Right[er] Way" to do these things.